style: Improve readability of land constructor
The previous constructor for lands was a function which just accept tons
of arguments into it.
Not only this reduced the readability, but also this is more error
prone

Instead, in this commit the arguments are passed as a table.
So it is obvious what value is passed to what argument and also there
won't be any nil arguments to the constructor (special lands for
example)